# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth Login & Protect API",
    "version": "1.0.0",
    "description": "A secure API using Supabase Auth for signup, login, logout, and protected routes."
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing or invalid fields" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user and return JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns access token" },
          "400": { "description": "Missing fields" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Terminate the user session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public, unprotected data",
        "responses": {
          "200": { "description": "Public info" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private user profile data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Read private dashboard data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard welcome message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    }
  }
}